+++ app/src/main/java/de/passwordvault/view/settings/dialog_license/LicenseDialog.java
@@ -2,13 +2,18 @@ package de.passwordvault.view.settings.dialog_license;
 
 import android.app.Dialog;
 import android.os.Bundle;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextClock;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.google.android.material.bottomsheet.BottomSheetBehavior;
+import com.google.android.material.bottomsheet.BottomSheetDialog;
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 
@@ -25,24 +30,36 @@ public class LicenseDialog extends BottomSheetDialogFragment {
     private LicenseViewModel viewModel;
 
 
-    @NonNull
+    @Nullable
     @Override
-    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         viewModel = new ViewModelProvider(this).get(LicenseViewModel.class);
-        MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(requireActivity());
-
         Bundle args = getArguments();
         if (args != null) {
             viewModel.processArguments(args);
         }
 
-        View view = getLayoutInflater().inflate(R.layout.dialog_license, null, false);
+        View view = inflater.inflate(R.layout.dialog_license, container, false);
+        ((TextView)view.findViewById(R.id.text_title)).setText(viewModel.getTitle());
         ((TextView)view.findViewById(R.id.text_license)).setText(viewModel.getLicense());
-        builder.setView(view);
-        builder.setTitle(viewModel.getTitle());
 
-
-        return builder.create();
+        return view;
     }
 
+
+    @NonNull
+    @Override
+    public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
+        if (getDialog() != null) {
+            getDialog().setOnShowListener(view -> {
+                BottomSheetDialog dialog = (BottomSheetDialog)view;
+                View bottomSheet = dialog.findViewById(com.google.android.material.R.id.design_bottom_sheet);
+                if (bottomSheet != null) {
+                    BottomSheetBehavior<View> behavior = BottomSheetBehavior.from(bottomSheet);
+                    behavior.setState(BottomSheetBehavior.STATE_COLLAPSED);
+                }
+            });
+        }
+        return super.onCreateDialog(savedInstanceState);
+    }
 }